move webpack config files to a folder
