+++ webpack/webpack.config.server.js
@@ -1,13 +1,14 @@
 import webpack from 'webpack';
 import ExtractTextPlugin from 'extract-text-webpack-plugin';
-import AssetsPlugin from 'assets-webpack-plugin';
 import path from 'path';
+import webpackNodeExternals from 'webpack-node-externals';
 import Dotenv from 'dotenv-webpack';
 import 'babel-polyfill';
 
 const GLOBALS = {
+  window: {},
   'process.env.NODE_ENV': JSON.stringify('production'),
-  'process.env.BROWSER': true,
+  'process.env.BROWSER': false,
   __DEV__: false
 };
 
@@ -16,31 +17,27 @@ export default {
     extensions: ['*', '.js', '.jsx', '.json']
   },
   devtool: 'source-map',
-  entry: ['babel-polyfill', path.resolve(__dirname, 'server/client')],
-  target: 'web',
+  entry: ['babel-polyfill', path.resolve(__dirname, '../server')],
+  target: 'node',
   mode: 'production',
   output: {
-    path: path.resolve(__dirname, 'server/build/public'),
+    path: path.resolve(__dirname, '../server/build'),
     publicPath: '/',
-    filename: 'client.js'
+    filename: 'server.js'
   },
   plugins: [
     new webpack.DefinePlugin(GLOBALS),
 
-    // Generate an external css file
     new ExtractTextPlugin('styles.css'),
 
     new Dotenv({
-      path: path.resolve(__dirname, `.env.${process.env.ENV || 'prod'}`),
+      path: path.resolve(__dirname, `../.env.${process.env.ENV || 'prod'}`),
       systemvars: true,
-    }),
-
-    // Output our JS and CSS files in a manifest file called assets.json
-    new AssetsPlugin({
-      path: path.resolve(__dirname, 'server/build'),
-      filename: 'assets.json',
     })
   ],
+  externals: [webpackNodeExternals({
+    whitelist: ['actioncable']
+  })],
   module: {
     rules: [
       { test: /\.jsx?$/, exclude: /node_modules/, loader: 'babel-loader' },
@@ -71,7 +68,7 @@ export default {
             }, {
               loader: 'sass-loader',
               options: {
-                includePaths: [path.resolve(__dirname, 'src', 'scss')],
+                includePaths: [path.resolve(__dirname, '../src', 'scss')],
                 sourceMap: true
               }
             }
